requirements.txt
streamlit
spektral
Pillow
tensorflow
networkx
matplotlib
seaborn
numpy
scikit-learn
scikit-image
opencv-python-headless
scipy
